menambah fitur peminjaman buku untuk user

+++ database/migrations/2025_03_24_171846_remove_description_from_books.php
@@ -9,12 +9,15 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
+        public function up()
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->dropColumn('description');
+            if (Schema::hasColumn('books', 'description')) {
+                $table->dropColumn('description');
+            }
         });
     }
+
     
     public function down()
     {
